automation_py/requirements.txt
pandas
tk
ghostscript
camelot-py
selenium
